move error boundary to errors folder
